# Optimize memory tracking with a `memThreshold`